chore: consolidate palette helpers and add phase5 verify docs
- Add shouldOpenPalette + stripSlashPrefix to palette.tsx (was missing)
- Delete duplicate palette-util.ts
- Update test imports to use ./palette
- Add docs/phase5-verify.md with flag verification steps
